--- court/src/db/mod.rs
@@ -1,1 +1,0 @@
-
