--- court/src/db/models.rs
@@ -1,10 +1,0 @@
-use diesel::prelude::*;
-
-#[derive(Queryable, Selectable)]
-#[diesel(table_name = crate::schema::recipes)]
-#[diesel(check_for_backend(diesel::pg::Pg))]
-pub struct Recipe {
-    pub id: i32,
-    pub title: String,
-    pub instructions: String,
-}
